+++ app/components/peopletable/peopletable.js
@@ -15,6 +15,7 @@
 * == BSD2 LICENSE ==
 */
 import React from 'react';
+import PropTypes from 'prop-types';
 import _ from 'lodash';
 import cx from 'classnames';
 import { Table, Column, Cell } from 'fixed-data-table-2';
@@ -26,6 +27,8 @@ import { translate, Trans } from 'react-i18next';
 import { SortHeaderCell, SortTypes } from './sortheadercell';
 import personUtils from '../../core/personutils';
 import ModalOverlay from '../modaloverlay';
+
+const resetSearchImageSrc = require('./images/searchReset.png');
 
 const TextCell = ({ rowIndex, data, col, icon, ...props }) => (
   <Cell {...props}>
@@ -39,24 +42,25 @@ const TextCell = ({ rowIndex, data, col, icon, ...props }) => (
 );
 
 TextCell.propTypes = {
-  col: React.PropTypes.string,
-  data: React.PropTypes.array,
-  rowIndex: React.PropTypes.number,
-  icon: React.PropTypes.object,
+  col: PropTypes.string,
+  data: PropTypes.array,
+  rowIndex: PropTypes.number,
+  icon: PropTypes.object,
 };
 
-const RemoveLinkCell = ({ rowIndex, data, handleClick, ...props }) => (
+const RemoveLinkCell = ({ rowIndex, data, handleClick, title, ...props }) => (
   <Cell {...props}>
     <div onClick={(e) => e.stopPropagation()} className="peopletable-cell remove">
-      <i onClick={handleClick(data[rowIndex], rowIndex)} className="peopletable-icon-remove icon-delete"></i>
+      <i title={title} onClick={handleClick(data[rowIndex], rowIndex)} className="peopletable-icon-remove icon-delete"></i>
     </div>
   </Cell>
 );
 
 RemoveLinkCell.propTypes = {
-  data: React.PropTypes.array,
-  rowIndex: React.PropTypes.number,
-  handleClick: React.PropTypes.func,
+  data: PropTypes.array,
+  rowIndex: PropTypes.number,
+  handleClick: PropTypes.func,
+  title: PropTypes.string.isRequired,
 };
 
 RemoveLinkCell.displayName = 'RemoveLinkCell';
@@ -72,10 +76,11 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
     this.handleRemovePatient = this.handleRemovePatient.bind(this);
     this.handleRowClick = this.handleRowClick.bind(this);
     this.handleSortChange = this.handleSortChange.bind(this);
-    this.handleToggleShowNames = this.handleToggleShowNames.bind(this);
     this.handleWindowResize = this.handleWindowResize.bind(this);
+    this.handleEmptySearch = this.handleEmptySearch.bind(this);
 
     this.state = {
+      searchPattern: '',
       currentRowIndex: -1,
       searching: false,
       showNames: true,
@@ -86,9 +91,14 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
       showModalOverlay: false,
       dialog: '',
       tableHeight: 590,
+      showSearchReset: false,
     };
 
     WindowSizeListener.DEBOUNCE_TIME = 50;
+  }
+
+  handleEmptySearch(e){
+    this.handleFilterChange(null);
   }
 
   componentDidMount() {
@@ -127,10 +137,12 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
   }
 
   handleFilterChange(e) {
-    if (_.isEmpty(e.target.value)) {
+    if (e === null ||_.isEmpty(e.target.value)) {
       this.setState({
         searching: false,
         dataList: this.buildDataList(),
+        showSearchReset: false,
+        searchPattern: '',
       });
       return;
     }
@@ -144,6 +156,8 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
     this.setState({
       searching: true,
       dataList: filtered,
+      showSearchReset: true,
+      searchPattern: e.target.value,
     });
   }
 
@@ -177,40 +191,24 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
         <div className="peopletable-search-label">
           {t('Patient List')}
         </div>
+        <div className="peopletable-search-box form-control-border">
         <input
           type="search"
-          className="peopletable-search-box form-control-border"
+          ref="searchInput"
+          className="input"
           onChange={this.handleFilterChange}
           placeholder={t('Search by Name')}
+          value={this.state.searchPattern}
         />
-      </div>
-    );
-  }
-
-  handleToggleShowNames() {
-
-    let toggleLabel = 'Clicked Hide All';
-    if ( !this.state.showNames ){
-      toggleLabel = 'Clicked Show All';
-    }
-
-    this.props.trackMetric(toggleLabel);
-    this.setState({ showNames: !this.state.showNames });
-  }
-
-  renderShowNamesToggle() {
-    const { t } = this.props;
-    let toggleLabel = t('Hide All');
-
-    if (!this.state.showNames) {
-      toggleLabel = t('Show All');
-    }
-
-    return (
-      <div className="peopletable-names-toggle-wrapper">
-        <a className="peopletable-names-toggle" disabled={this.state.searching} onClick={this.handleToggleShowNames}>
-          {toggleLabel}
-        </a>
+        { 
+        this.state.showSearchReset ? 
+          <img
+            onClick={this.handleEmptySearch}
+            className="peopletable-reset-image"
+            src={resetSearchImageSrc}
+            title={t('Reset Search')}/> : null 
+        }
+        </div>
       </div>
     );
   }
@@ -224,14 +222,6 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
   handleRowClick(e, rowIndex) {
     this.props.trackMetric('Selected PwD');
     browserHistory.push(this.state.dataList[rowIndex].link);
-  }
-
-  renderPeopleInstructions() {
-    return (
-      <Trans className="peopletable-instructions" i18nKey="html.peopletable-instructions">
-        Type a patient name in the search box or click <a className="peopletable-names-showall" onClick={this.handleToggleShowNames}>Show All</a> to display all patients.
-      </Trans>
-    );
   }
 
   renderRemoveDialog(patient) {
@@ -320,6 +310,8 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
     const { t } = this.props;
     const { colSortDirs, dataList, tableWidth, tableHeight } = this.state;
 
+    const title = t('I want to quit this patient\'s care team');
+
     return (
       <Table
         rowHeight={65}
@@ -374,6 +366,7 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
           cell={<RemoveLinkCell
             data={dataList}
             handleClick={this.handleRemove.bind(this)}
+            title={title}
           />}
           width={30}
           flexGrow={0}
@@ -382,22 +375,11 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
     )
   }
 
-  renderPeopleArea() {
-    const { showNames, searching } = this.state;
-
-    if (!showNames && !searching) {
-      return this.renderPeopleInstructions();
-    } else {
-      return this.renderPeopleTable();
-    }
-  }
-
   render() {
     return (
       <div>
         {this.renderSearchBar()}
-        {this.renderShowNamesToggle()}
-        {this.renderPeopleArea()}
+        {this.renderPeopleTable()}
         {this.renderModalOverlay()}
         <WindowSizeListener onResize={this.handleWindowResize} />
       </div>
@@ -406,9 +388,9 @@ const PeopleTable = translate()(class PeopleTable extends React.Component {
 });
 
 PeopleTable.propTypes = {
-  people: React.PropTypes.array,
-  trackMetric: React.PropTypes.func.isRequired,
-  onRemovePatient: React.PropTypes.func.isRequired,
+  people: PropTypes.array,
+  trackMetric: PropTypes.func.isRequired,
+  onRemovePatient: PropTypes.func.isRequired,
 };
 
-module.exports = PeopleTable;
+export default PeopleTable;