+++ CHANGELOG.md
@@ -12,6 +12,7 @@ It is based on Tidepool Blip 1.27.
 - PT-1052 Rework sign up (direct link to clinician signup from home page + clinician option selected as default in signup page)
 - PT-1100 Move message notes to the bottom of the daily page
 - PT-1041 Rework invitation page for clinicals users
+- PT-1111 Allow capacity to reset the search
 
 ### Added
 - PT-875 Flexible period in Trending page